LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/fill-and-stroke-styles/2d.fillStyle.parse.hsl-clamp-2-expected.txt

2d.fillStyle.parse.hsl-clamp-2


FAIL OffscreenCanvas test: 2d.fillStyle.parse.hsl-clamp-2 assert_equals: Red channel of the pixel at (50, 25) expected 127 but got 255
